refactor: move component tests to separate file
