+++ packages/integrations/markdoc/test/render.test.js
@@ -47,42 +47,6 @@ describe('Markdoc - render', () => {
 			await server.stop();
 		});
 
-		it('renders content - with components', async () => {
-			const fixture = await getFixture('render-with-components');
-			const server = await fixture.startDevServer();
-
-			const res = await fixture.fetch('/');
-			const html = await res.text();
-
-			renderComponentsChecks(html);
-
-			await server.stop();
-		});
-
-		it('renders content - with components inside partials', async () => {
-			const fixture = await getFixture('render-with-components');
-			const server = await fixture.startDevServer();
-
-			const res = await fixture.fetch('/');
-			const html = await res.text();
-
-			renderComponentsInsidePartialsChecks(html);
-
-			await server.stop();
-		});
-
-		it('renders content - with indented components', async () => {
-			const fixture = await getFixture('render-with-indented-components');
-			const server = await fixture.startDevServer();
-
-			const res = await fixture.fetch('/');
-			const html = await res.text();
-
-			renderIndentedComponentsChecks(html);
-
-			await server.stop();
-		});
-
 		it('renders content - with `render: null` in document', async () => {
 			const fixture = await getFixture('render-null');
 			const server = await fixture.startDevServer();
@@ -136,33 +100,6 @@ describe('Markdoc - render', () => {
 			renderConfigChecks(html);
 		});
 
-		it('renders content - with components', async () => {
-			const fixture = await getFixture('render-with-components');
-			await fixture.build();
-
-			const html = await fixture.readFile('/index.html');
-
-			renderComponentsChecks(html);
-		});
-
-		it('renders content - with components inside partials', async () => {
-			const fixture = await getFixture('render-with-components');
-			await fixture.build();
-
-			const html = await fixture.readFile('/index.html');
-
-			renderComponentsInsidePartialsChecks(html);
-		});
-
-		it('renders content - with indented components', async () => {
-			const fixture = await getFixture('render-with-indented-components');
-			await fixture.build();
-
-			const html = await fixture.readFile('/index.html');
-
-			renderIndentedComponentsChecks(html);
-		});
-
 		it('renders content - with `render: null` in document', async () => {
 			const fixture = await getFixture('render-null');
 			await fixture.build();
@@ -194,35 +131,6 @@ function renderNullChecks(html) {
 }
 
 /** @param {string} html */
-function renderComponentsChecks(html) {
-	const { document } = parseHTML(html);
-	const h2 = document.querySelector('h2');
-	assert.equal(h2.textContent, 'Post with components');
-
-	// Renders custom shortcode component
-	const marquee = document.querySelector('marquee');
-	assert.notEqual(marquee, null);
-	assert.equal(marquee.hasAttribute('data-custom-marquee'), true);
-
-	// Renders Astro Code component
-	const pre = document.querySelector('pre');
-	assert.notEqual(pre, null);
-	assert.equal(pre.className, 'astro-code github-dark');
-}
-
-/** @param {string} html */
-function renderComponentsInsidePartialsChecks(html) {
-	const { document } = parseHTML(html);
-	// renders Counter.tsx
-	const button = document.querySelector('#counter');
-	assert.equal(button.textContent, '1');
-
-	// renders DeeplyNested.astro
-	const deeplyNested = document.querySelector('#deeply-nested');
-	assert.equal(deeplyNested.textContent, 'Deeply nested partial');
-}
-
-/** @param {string} html */
 function renderPartialsChecks(html) {
 	const { document } = parseHTML(html);
 	const top = document.querySelector('#top');
@@ -231,26 +139,6 @@ function renderPartialsChecks(html) {
 	assert.ok(nested);
 	const configured = document.querySelector('#configured');
 	assert.ok(configured);
-}
-
-/** @param {string} html */
-function renderIndentedComponentsChecks(html) {
-	const { document } = parseHTML(html);
-	const h2 = document.querySelector('h2');
-	assert.equal(h2.textContent, 'Post with indented components');
-
-	// Renders custom shortcode components
-	const marquees = document.querySelectorAll('marquee');
-	assert.equal(marquees.length, 2);
-
-	// Renders h3
-	const h3 = document.querySelector('h3');
-	assert.equal(h3.textContent, 'I am an h3!');
-
-	// Renders Astro Code component
-	const pre = document.querySelector('pre');
-	assert.notEqual(pre, null);
-	assert.equal(pre.className, 'astro-code github-dark');
 }
 
 /** @param {string} html */
